simplify metric
