+++ src/pie_modules/taskmodules/metrics/__init__.py
@@ -1,1 +1,1 @@
-from .rouge import RougeMetric
+from .text import TextMetric
